+++ rentacar/src/main/java/com/oguzhanturk/rentacar/business/dtos/rental/ListRentalDto.java
@@ -19,8 +19,8 @@ public class ListRentalDto {
 	private String fromCityName;
 	private String toCityName;
 	private List<String> additionalServiceName;
-	private BigDecimal rentalDailyPrice;
-	private BigDecimal rentalTotalPrice;
+	private double rentalDailyPrice;
+	private double rentalTotalPrice;
 	private int customerId;
 	private int carCarId;
