+++ rentacar/src/main/java/com/oguzhanturk/rentacar/business/dtos/rental/RentalDto.java
@@ -21,8 +21,8 @@ public class RentalDto {
 	private String fromCityName;
 	private String toCityName;
 	private List<String> additionalServiceName;
-	private BigDecimal rentalDailyPrice;
-	private BigDecimal rentalTotalPrice;
+	private double rentalDailyPrice;
+	private double rentalTotalPrice;
 	private int customerId;
 	private int carCarId;
