+++ rentacar/src/main/java/com/oguzhanturk/rentacar/business/request/additionalService/CreateAdditionalServiceRequest.java
@@ -13,5 +13,5 @@ public class CreateAdditionalServiceRequest {
 
 	private String additionalServiceName;
 	private String additionalServiceDescription;
-	private BigDecimal additionalServiceDailyPrice;
+	private double additionalServiceDailyPrice;
 }
