+++ rentacar/src/main/java/com/oguzhanturk/rentacar/business/request/additionalService/UpdateAdditionalServiceRequest.java
@@ -14,5 +14,5 @@ public class UpdateAdditionalServiceRequest {
 	private int additionalId;
 	private String additionalServiceName;
 	private String additionalServiceDescription;
-	private BigDecimal additionalServiceDailyPrice;
+	private double additionalServiceDailyPrice;
 }
