+++ rentacar/src/main/java/com/oguzhanturk/rentacar/entities/concretes/Car.java
@@ -24,39 +24,39 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Car {
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "car_id")
-	private int carId;
-	@Column(name = "daily_price")
-	private BigDecimal dailyPrice;
-	@Column(name = "model_year")
-	private int modelYear;
-	@Column(name = "description")
-	private String description;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "car_id")
+    private int carId;
+    @Column(name = "daily_price")
+    private double dailyPrice;
+    @Column(name = "model_year")
+    private int modelYear;
+    @Column(name = "description")
+    private String description;
 
-	@ManyToOne
-	@JoinColumn(name = "brand_id")
-	private Brand brand;
+    @ManyToOne
+    @JoinColumn(name = "brand_id")
+    private Brand brand;
 
-	@ManyToOne
-	@JoinColumn(name = "color_id")
-	private Color color;
+    @ManyToOne
+    @JoinColumn(name = "color_id")
+    private Color color;
 
-	@Column(name = "kilometer")
-	private double kilometer;
+    @Column(name = "kilometer")
+    private double kilometer;
 
-	@ManyToOne
-	@JoinColumn(name = "city_id")
-	private City city;
+    @ManyToOne
+    @JoinColumn(name = "city_id")
+    private City city;
 
-	@OneToMany(mappedBy = "car")
-	private List<CarMaintenance> carMaintenances;
+    @OneToMany(mappedBy = "car")
+    private List<CarMaintenance> carMaintenances;
 
-	@OneToMany(mappedBy = "car")
-	private List<Rental> rentals;
+    @OneToMany(mappedBy = "car")
+    private List<Rental> rentals;
 
-	@OneToMany(mappedBy = "car")
-	private List<CarAccident> carAccidents;
+    @OneToMany(mappedBy = "car")
+    private List<CarAccident> carAccidents;
 
 }
